Added ros topic that publishes all mentions of the robot's account.

+++ twitcher_connection/include/twitcher_connection/TwitterMentions.h
@@ -27,7 +27,7 @@
 class TwitterMentions : public TwitterApiCall
 {
 public:
-    TwitterMentions(std::string configFile, int count, int since_id, int max_id,
+    TwitterMentions(std::string configFile, int count, long since_id, long max_id,
                     bool trim_user, bool contributor_details,
                     bool include_entities);
     
@@ -37,8 +37,8 @@ public:
     
 private:
     int count;
-    int since_id;
-    int max_id;
+    long since_id;
+    long max_id;
     bool trim_user;
     bool contributor_details;
     bool include_entities;
